Fix error when displaying the status of collect quests
* Remove the call to the 'count' key in the 'collect' dictionary

It appears that the data structure for 'collect quests' changed, and so
printing the progress throws an error. Before, the first value of the 'collect'
dictionary referenced another dictionary with a 'count' key.  Now the 'collect'
value has a dictionary with the key being the name of the object that is being
collected and the value is the number of gathered items. Thus, removing the call
to the 'count' key solved the problem.